fix: renaming env package
