+++ traverse/entry.go
